package socket.io behavior
